プリファレンス修正。ヘルプ修正。

+++ src/com/yochiyochi/android/UntouchableTimer/TimerService.java
@@ -17,7 +17,7 @@ public class TimerService extends Service {
     {  Log.d(TAG, "@@@---start---@@@"); }
 
 	Context mContext;
-	int counter;
+	long counter;
 	Timer timer;
 	public PowerManager.WakeLock wl;
